Add add edit popup model, WIP
